# backend/src/docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "innov8 Backend API",
    "version": "1.0.0",
    "description": "OpenAPI coverage for the live innov8 backend routes under `backend/src/routes`.\n\nAuthentication flow:\n1. Call `POST /auth/challenge` with a Stellar public key.\n2. Sign the returned challenge with the wallet.\n3. Exchange the signature at `POST /auth/verify` for a bearer token.\n4. Send `Authorization: Bearer <token>` to protected endpoints.\n\nJWT semantics enforced by middleware:\n- `iss` must match `JWT_ISSUER` (default `amana`)\n- `aud` must match `JWT_AUDIENCE` (default `amana-api`)\n- `jti` is required and checked against the revocation denylist\n  (revoked tokens are rejected — replay protection)\n- `iat` (issued-at) must be present and in the past\n- `exp` (expires-at) is enforced: expired tokens are rejected\n  (default TTL: 86400 s / 24 h, configurable via `JWT_EXPIRES_IN`)\n- `nbf` (not-before) must be present and not be in the future\n- Tokens signed with algorithms other than HS256 are rejected\n\nError code references in operation descriptions map to backend `ErrorCode` values in\n`backend/src/errors/errorCodes.ts`. Some route handlers still emit legacy `{ error }`\npayloads while centralized validation/runtime failures return `{ code, message, details }`."
  },
  "servers": [
    {
      "url": "http://localhost:4000"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "adminAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-admin-secret",
        "description": "Secret key authentication for admin-only management routes."
      }
    },
    "parameters": {
      "TradeIdPath": {
        "in": "path",
        "name": "id",
        "required": true,
        "description": "Trade identifier used by the backend route layer.",
        "schema": {
          "type": "string"
        },
        "example": 4294967297
      },
...
      "IdempotencyKeyHeader": {
        "in": "header",
        "name": "Idempotency-Key",
        "required": false,
        "description": "Optional idempotency key cached for mutation requests for up to 24 hours.",
        "schema": {
          "type": "string"
        }
      }
    },
    "schemas": {
      "ErrorEnvelope": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ],
        "example": {
          "error": "Unauthorized"
        }
      },
      "AppErrorResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "enum": [
              "VALIDATION_ERROR",
              "AUTH_ERROR",
              "DOMAIN_ERROR",
              "INFRA_ERROR",
              "NOT_FOUND",
              "INTERNAL_ERROR",
              "CLAWBACK_TOO_LARGE",
              "CLAWBACK_INVALID_AMOUNT"
            ]
          },
          "message": {
            "type": "string"
          },
          "details": {
            "oneOf": [
              {
                "type": "object"
              },
              {
                "type": "array",
                "items": {
                  "type": "object"
                }
              }
            ]
          }
        },
        "required": [
          "code",
          "message",
          "details"
        ],
        "example": {
          "code": "VALIDATION_ERROR",
          "message": "Validation failed",
          "details": [
            {
              "path": "body.amountUsdc",
              "message": "Invalid amount format"
            }
          ]
        }
      },
...
      "UnsignedXdrResponse": {
        "type": "object",
        "properties": {
          "unsignedXdr": {
            "type": "string"
          }
        },
        "required": [
          "unsignedXdr"
        ]
      },
...
    },
    "responses": {
      "UnauthorizedError": {
        "description": "Missing, invalid, revoked, or semantically invalid bearer token.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            },
            "examples": {
              "unauthorized": {
                "value": {
                  "error": "Unauthorized"
                }
              },
              "revoked": {
                "value": {
                  "error": "Unauthorized: token has been revoked"
                }
              }
            }
          }
        }
      },
      "ForbiddenError": {
        "description": "Authenticated caller is not allowed to perform the action.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            },
            "example": {
              "error": "Forbidden"
            }
          }
        }
      },
      "NotFoundError": {
        "description": "Requested resource was not found.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            },
            "example": {
              "error": "Trade not found"
            }
          }
        }
      },
      "LegacyValidationError": {
        "description": "Request payload failed route-level validation.",
        "content": {
          "application/json": {
            "schema": {
              "oneOf": [
                {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                },
                {
                  "$ref": "#/components/schemas/AppErrorResponse"
                }
              ]
            }
          }
        }
      },
      "InternalError": {
        "description": "Unexpected backend failure.",
        "content": {
          "application/json": {
            "schema": {
              "oneOf": [
                {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                },
                {
                  "$ref": "#/components/schemas/AppErrorResponse"
                }
              ]
            }
          }
        }
      }
    }
  },
  "paths": {
...
    "/api/admin/payouts": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "List payout intents",
        "description": "List payout intents newest first so admins can find stuck or failed\npayouts without querying the database. Filter by `status` and by an\ninclusive `from`/`to` range on `createdAt`. Uses opaque cursor-based\npagination (`cursor` / `pageInfo.nextCursor`).",
        "security": [
          {
            "bearerAuth": [],
            "adminAuth": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "PENDING",
                "SUBMITTED",
                "CONFIRMED",
                "FAILED"
              ]
            }
          },
          {
            "name": "from",
            "in": "query",
            "description": "Only intents created at or after this timestamp.",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "to",
            "in": "query",
            "description": "Only intents created at or before this timestamp.",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "description": "Opaque cursor from a previous response's `pageInfo.nextCursor`.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A page of payout intents.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "items",
                    "pageInfo"
                  ],
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "integer"
                          },
                          "idempotencyKey": {
                            "type": "string"
                          },
                          "kind": {
                            "type": "string"
                          },
                          "tradeId": {
                            "type": "string"
                          },
                          "milestoneIndex": {
                            "type": "integer",
                            "nullable": true
                          },
                          "amountUsdc": {
                            "type": "string"
                          },
                          "destination": {
                            "type": "string"
                          },
                          "requestedBy": {
                            "type": "string"
                          },
                          "status": {
                            "type": "string",
                            "enum": [
                              "PENDING",
                              "SUBMITTED",
                              "CONFIRMED",
                              "FAILED"
                            ]
                          },
                          "txHash": {
                            "type": "string",
                            "nullable": true
                          },
                          "duplicateAttempts": {
                            "type": "integer"
                          },
                          "lastError": {
                            "type": "string",
                            "nullable": true
                          },
                          "submittedAt": {
                            "type": "string",
                            "format": "date-time",
                            "nullable": true
                          },
                          "confirmedAt": {
                            "type": "string",
                            "format": "date-time",
                            "nullable": true
                          },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "updatedAt": {
                            "type": "string",
                            "format": "date-time"
                          }
                        }
                      }
                    },
                    "pageInfo": {
                      "type": "object",
                      "properties": {
                        "nextCursor": {
                          "type": "string",
                          "nullable": true
                        },
                        "hasNextPage": {
                          "type": "boolean"
                        },
                        "limit": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid filter, limit or cursor."
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          }
        }
      }
    },
...
    "/trades/{id}/release": {
      "post": {
        "tags": [
          "Trades"
        ],
        "summary": "Build a release-funds transaction",
        "description": "Allowed for the buyer or an admin wallet only, and the trade must be `DELIVERED`.\nSupports optional `Idempotency-Key`.\nError code references: `AUTH_ERROR`, `VALIDATION_ERROR`, `DOMAIN_ERROR`, `NOT_FOUND`, `INTERNAL_ERROR`.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/TradeIdPath"
          },
          {
            "$ref": "#/components/parameters/IdempotencyKeyHeader"
          }
        ],
        "responses": {
          "200": {
            "description": "Unsigned release transaction returned.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnsignedXdrResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/LegacyValidationError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "description": "Caller is not allowed to release funds.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                },
                "examples": {
                  "buyerOrAdmin": {
                    "value": {
                      "error": "Only the buyer or an admin may release funds"
                    }
                  },
                  "generic": {
                    "value": {
                      "error": "Forbidden"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        },
        "operationId": "post.trades.id.release"
      }
    },
...
  }
}